A typo error on my computer
Traceback (most recent call last):
  File "main.py", line 64, in <module>
    main(config)
  File "main.py", line 33, in main
    solver.train()
  File "/home/user/PycharmProjects/SpeechSplit-master/solver.py", line 161, in train
    x_f0_intrp = self.Interp(x_f0, len_org)
  File "/home/user/anaconda3/envs/vencen/lib/python3.6/site-packages/torch/nn/modules/module.py", line 547, in __call__
    result = self.forward(*input, **kwargs)
  File "/home/user/PycharmProjects/SpeechSplit-master/model.py", line 407, in forward
    idx_mask = idx_scaled_fl < (len_seg - 1)
RuntimeError: Expected object of scalar type Float but got scalar type Long for argument #2 'other'
Aborted (core dumped)
exit status 134

+++ model.py
@@ -402,7 +402,7 @@ class InterpLnr(nn.Module):
                                 device=device)
         
         # end point of each segment
-        idx_mask = idx_scaled_fl < (len_seg - 1)
+        idx_mask = idx_scaled_fl < (len_seg - 1).float()
        
         offset = len_seg.view(batch_size, -1).cumsum(dim=-1)
         # offset starts from the 2nd segment
